Changed ssize_t for long
On 32-bits OS the ssize_t is the same size of a int and not a long

+++ C00/ex07/ft_putnbr.c
@@ -19,7 +19,7 @@
 void	ft_putnbr(int nb)
 {
 	char	c;
-	ssize_t	n;
+	long	n;
 
 	n = nb;
 	if (n < 0)
